Created view object, renamed methods, cleaned up
In this version: 
A new App.update method takes the place of the old App.render and is responsible for saving data to storage. 
A new view object now contains display-only methods .render and .renderFooter.
The editKeyup is renamed editHighlight.
The old .update method is now renamed editAction.
All corresponding methods() are called appropriately.